[REFACTOR] runners, logs, sa
